Use usize for `num_values`

+++ src/arrow2/src/io/parquet/read/deserialize/binary/nested.rs
@@ -144,7 +144,7 @@ pub struct NestedIter<O: Offset, I: Pages> {
     dict: Option<Dict>,
     chunk_size: Option<usize>,
     rows_remaining: usize,
-    values_remaining: i64,
+    values_remaining: usize,
 }
 
 impl<O: Offset, I: Pages> NestedIter<O, I> {
@@ -154,7 +154,7 @@ impl<O: Offset, I: Pages> NestedIter<O, I> {
         data_type: DataType,
         num_rows: usize,
         chunk_size: Option<usize>,
-        num_values: i64,
+        num_values: usize,
     ) -> Self {
         Self {
             iter,
